Don't run jobs if a runlock already exists
We have an issue where the same job is being run on different workers on
the same time.

We know this presents a problem if one of our workers dies hard whilst
running a job. The runlock will be deleted once we go over the TTL and
attempt to enqueue another job though.